refactor: 후기 작성 관련 파일들 common폴더로 이동
