--- src/pages/landlord/repair/L_WriteReview.jsx
@@ -1,9 +1,0 @@
-import L_WriteReviewTemplate from '../../../templates/landlord/repair/L_WriteReviewTemplate';
-
-export default function WriteReview() {
-  return (
-    <div>
-      <L_WriteReviewTemplate />
-    </div>
-  );
-}
